// src/models/interface/categories/Newscategory.ts
export interface NewsCategory {
    id_detalle: number
    numero_categoria: number
    nombre_categoria: string
    codigo_categoria: string
    descripcion: string
    alias: string
    estado: string
    id_usuario_creador: number
    nombre_usuario_creador: string
    nombre_usuario_actualizador: any
    fecha_creado: string
    categoria_general: string
  }